:recycle: Refactor: Test Api 위치 이동
<footer>
- 관련: #139
