+++ application/external-api/src/main/java/com/namo/spring/application/external/api/test/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.namo.spring.application.external.domain.test.ui;
+package com.namo.spring.application.external.api.test.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.namo.spring.application.external.domain.test.ui.dto.TestRequest;
-import com.namo.spring.application.external.domain.test.ui.dto.TestResponse;
+import com.namo.spring.application.external.api.test.dto.TestRequest;
+import com.namo.spring.application.external.api.test.dto.TestResponse;
 import com.namo.spring.application.external.global.annotation.swagger.ApiErrorCode;
 import com.namo.spring.application.external.global.annotation.swagger.ApiErrorCodes;
 import com.namo.spring.core.common.code.status.ErrorStatus;
